use std::lerp, constexpr hex.h
