+++ src/Backups/IBackupCoordination.h
@@ -2,7 +2,7 @@
 
 #include <optional>
 #include <fmt/format.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <Core/Types.h>
