+++ src/Common/BinStringDecodeHelper.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Common/hex.h>
+#include <base/hex.h>
 
 namespace DB
 {
