+++ src/Common/OpenTelemetryTraceContext.cpp
@@ -3,7 +3,7 @@
 #include <random>
 #include <base/getThreadId.h>
 #include <Common/Exception.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <Core/Settings.h>
 #include <IO/Operators.h>
