+++ src/Storages/Cache/ExternalDataSourceCache.cpp
@@ -15,7 +15,7 @@
 #include <Common/Exception.h>
 #include <Common/ProfileEvents.h>
 #include <Common/SipHash.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <Core/Types.h>
 #include <base/types.h>
 #include <consistent_hashing.h>
