+++ src/Storages/Distributed/DistributedAsyncInsertDirectoryQueue.cpp
@@ -17,7 +17,7 @@
 #include <Common/StringUtils/StringUtils.h>
 #include <Common/SipHash.h>
 #include <Common/quoteString.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 #include <Common/ActionBlocker.h>
 #include <Common/formatReadable.h>
 #include <Common/Stopwatch.h>
