+++ src/Storages/LiveView/LiveViewSink.h
@@ -3,7 +3,7 @@
 #include <DataTypes/DataTypesNumber.h>
 #include <Processors/Sinks/SinkToStorage.h>
 #include <Storages/LiveView/StorageLiveView.h>
-#include <Common/hex.h>
+#include <base/hex.h>
 
 
 namespace DB
